+++ lib/legion/api.rb
@@ -42,6 +42,7 @@ require_relative 'api/governance'
 require_relative 'api/acp'
 require_relative 'api/prompts'
 require_relative 'api/marketplace'
+require_relative 'api/apollo'
 require_relative 'api/graphql' if defined?(GraphQL)
 
 module Legion
@@ -129,6 +130,7 @@ module Legion
     register Routes::Acp
     register Routes::Prompts
     register Routes::Marketplace
+    register Routes::Apollo
     register Routes::GraphQL if defined?(Routes::GraphQL)
 
     use Legion::API::Middleware::RequestLogger